ingress-nginx: Pass X-Accel-Buffering header through proxy
Allow X-Accel-Buffering nginx control headers to pass through the nginx
ingress proxy. An "X-Accel-Buffering: no" header can be sent from
a proxied webservice to tell nginx not to buffer its response to the
client. This header however is stripped by the nginx http proxy module's
default config. For the header to also reach the downstream Toolforge
front proxy we must explicitly allow it to be propagated.

Bug: T363877