jobs-api: reflect production deployment
I messed up here:
https://gitlab.wikimedia.org/repos/cloud/toolforge/toolforge-deploy/-/merge_requests/289

when I forgot to merge the MR and the MR creator bot added on the same
MR the changes for bumping to 0.0.303, and without noticing then I
merged the MR with the newer changes to.

Then the revert:
https://gitlab.wikimedia.org/repos/cloud/toolforge/toolforge-deploy/-/merge_requests/290

Undid both changes. This commit sets the current status of deployment
that is 0.0.302.

Signed-off-by: David Caro <dcaro@wikimedia.org>