kyverno: introduce configuration to support our cluster
Previous to this patch, we had disabled all resource limits for kyverno.
We believe this was part of the reason we suffered a kubernetes-wide outage.

Introduce some resource limits, which are otherwise pretty permissive.

While at it, upstream recommended disabling some of the components that are kind
of optional, and can cause additional load.

They also recommended increasing the number of replicas for the admission controller.

Bug: T367386
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>