0.0.416-update-resources.sh: add migration script
It updates all the resources with default limits to the new defaults.
Note that it will force deployments to restart their pods.

It does not change any one-offs, only scheduled and continuous jobs.

It includes a dry-run mode that will allow to test without changing
anything in tools with a big number of namespaces.

Signed-off-by: David Caro <dcaro@wikimedia.org>