loki.alloy: don't drop logs and increase rate limit
As we are fetching logs from files, and those files might have a high
amount of logs, we are hitting the rate limit really often, almost
every time that it starts tailing a new file. That currently drops any
lines exceeding this limit making logs incomplete.

This prevents this by not dropping any logs, and instead just
rate-limiting.

Note that it's not per-tool, as you have to set the drop policy if
you want to apply limits per-tool.

We might want to follow up with more tweaks or memory limiting options
if we see alloy is not able to keep up with the log rate.

Bug: T402736\
Signed-off-by: David Caro <dcaro@wikimedia.org>