components-api - drop redundant config calls
We call `flush_everything` during the setup, so the first config create
has a clean environment.

The following config create calls will perform an update, we don't need
to explicitly delete it first.

This saves a number of round-trips to the API and follows the existing
pattern of re-using previous test state.