deploy-tests - no flush if skipping
Currently we flush all resources in this test suite, even when all tests
will be skipped due to a prior failure, such as in `components-smoke-test`.

This makes it very hard to debug why a previous test failed (you have to
"catch it in the act" as the state has been deleted after the failure).

Guard the flush with the same logic we use for skip, so the resources
will be left dangling in cause of an earlier failure.