run_functional_tests - abort on missing branch
The default branch is already set to `main` in the entrypoint, so the
branch check will either fail for a user specified branch, or the
default.

In case it is the default, then setting to the default, is potentially
buggy (branch rename).

In case the user has specified a branch, then switching the target is
unexpected behaviour.

Given the verbosity of output it is very easy to miss the line notifying
the branch has been switched.

Change the logic to explicitly fail if the branch is not found.