components-smoke-test - explicitly write config
When bats is executed with `--verbose-run`, the output from `run` is
emitted from the call to the standard FDs.

Within the config generation tests, we rely on the stdout to write the
config file. Without `--verbose-run` this file is blank and fails
validation during loading.

Bats guarantees that `$output` is set with the contents of stdout, so
rather than using redirection from the `run`, write `$output` explicitly
after verifying the command was successfuly.